wip: close bomb progression with bomb planted

+++ Source/MyProject/Components/C_Interactive.h
@@ -69,7 +69,7 @@ private:
 	bool bInteracting;
 
 	UPROPERTY(VisibleAnywhere, Replicated)
-	double InteractionStartWorldTime;
+	float InteractionStartWorldTime;
 
 	UPROPERTY(VisibleAnywhere, Replicated)
 	AA_Character* Interactor;
